CORE-14869 & CORE-14872 Fix retry of initial registration (#4142)
Retrying an initial registration after it had been declined was not working since the MGM created a pending member info when it processed the initial registration the first time and the logic in place assumed that during the second attempt, the name was already taken by a different member.
This change should allow a new registration to go through if the previous version of the member info was never approved.

An additional problem was found after resolving that which was that the member signature verification failed after member list distribution. This was because the persisted member info has non updateable fields around the member signatures and so the old member signatures were stored after persisting pending member info again for the same serial number. These are being switched to updateable now in line with what has been done in 5.1 as part of the re-registration work.