CORE-15068 Reduce the external event retry period (#4184)
The external event retry period relies on other events to come in after
the retry period has elapsed.

Due to the increased duration between flow wakeup events, caused flows
that have to retry external events to wait a significantly long time to
retry.

When an external event must be retried, it will schedule a wakeup
that matches the `sendTimestamp` set when sending the previous attempt
for the external event. Therefore, decoupling the retry period from the
general wakeup period of flows.