CORE-14658: stop using lambda in cipher suite metrics due to Quasar problems, use our own PEM parsing logic which cannot block (#4215)
Quasar seems to have trouble with lambdas in the way we use metrics, and that caused misbehaviour. To be safe, avoid lambdas and remove the use of theoretically blocking BufferReader when parsing PEM content. This involves doing the top level parse of PEM content ourselves, which also allows detailed error reporting than the Bouncy Castle code. The detailed and security sensitive ASN.1 handling is still done by Bouncy Castle. With these changes, we no longer see the fiber corruption (e.g null pointers out of nowhere, lines of code being skipped) we were seeing, but since the issue is very sensitive to timing it's hard to be sure.

Add more test cases for PEM parsing now we own the logic.