CORE-14709: Adding apiVersion dimension to REST API - fix test exception when introducing new endpoint (#4248)
After changes introduced in CORE-14709, when a baseline was not updated with a new endpoint OpenApiCompatibilityTest threw a runtime exception `Index out of bounds` and it didn't produce an error report and new baseline. Fix - after detecting differences in actual and expected number of tags don't perform each tag pair comparision.