CORE-15278, CORE-10051:CORE-15278, CORE-10051: remove unused types and code that handle situations where the private key was stored in external HSM (multi-HSM concept), since we are not supporting multiple HSMs (#4242)
CORE-14335: make SigningKeyInfo.keyMaterial non-nullable

SiginingKeyInfo.keyMaterial used to be nullable, in order to support HSMs where we do not have the key material (i.e. we do not have access to an encrypted private key). Tlet's remove support for missing key material in signing key records since it can never happen in production right now with our current feature set. This is a step towards removing the HSM layers in order to squash together code in order to improve performance.

CORE-14335: always pass SigningWrappedSpec into CryptoService.sign, so delete the interface and SigningAliasSpec which is the other implementation class we never use (#4075)

Currently CryptoService.sign took a SigningSpec as an argument, which is an interface which had 2 implementations; the method is polymorphic. One variant, SigningAliasSpec, was intended for use when signing when we only know the alias of a key, but not the private key itself. This is not supported, and we do not want to carry the complexity.

Therefore, delete the interface and SigningAliasSpec which is an implementation class we never use, since it was part of the multiple HSM support. This change will make it easier to improve the caching behaviour of the crypto processor, since the current complexity we are addressing here means there can be multiple crypto services, which brings complexity to the fast path.

I considered making SigningWrappedSpec be called SigningSpec. I felt that would potentially be more confusing at this stage, but that's a marginal call. Long run, a lot of the Spec classes in the crypto processor can be replaced by a small number of arguments to various functions.

This is a step toward CORE-10051.

Use check rather than require

Remove nullability on master wrapping key

Remove GeneratedPublicKey and SigningPublicKeySaveContext