Show the x509 name of the first certificate from the chain if cert verification fails (#4279)
The signature of the notary CPB now includes the certificate chain.
We were relying on the fact that if the validation of the certificate fails, it shows the index of the certificate which failed the validation. Unfortunately the documentation says, it MAY include this index, but it doesn't have to. So we can't rely on that information.
We now instead always print the X509 name of the first certificate in the chain if the index is not set.