CORE-15354 - Improve reason for failed registration in case of timeout (#4280)
In cases where a member didn't get any update from the MGM for a registration request, the reason was including the full registration request that was then getting cropped, so it wasn't clear what the reason was. Cleaned up the message and clarified it a bit to make it a bit easier for the user to understand.