CORE-10051: rename HSMServiceImpl and rename to TenantInfoService, without OSGi; eliminate HSMStore and CryptoServiceFactory (#4253)
CORE-10051: remove HSMServiceImpl OSGi, and rename to TenantInfoService, since the name is misleading because it is not really anything to do with HSMs, since we currently don't support HSMs .  Unfortunately this means the names in the code don't match the names in the public API (DB schema and Avro messages), but that's fairly  localised and encapsulated in the crypto code so hopefully this is a good tradeoff. 

Remove HSMStore; use HSMRepository directly instead, to remove a layer of code.

Remove hsmId as far as possible, since it only ever has one value ("SOFT").

Move CryptoService to from libs/crypto/cipher-suite to libs/crypto/crypto-core so that it can use declarations in crypto-core.

Use a string constant for SOFT HSM rather than repeating string literal in various tests and other places.

Optimise imports.

Remove unused constants.