CORE-9519 API to change group parameters (#4284)
Adds an endpoint to the MGM API to update group parameters. The MGM may update the current group parameters by submitting a map of the updated version (which may only include changes to the minimum platform version and custom ("ext.") properties). The platform populates notary information to construct the final group parameters, which are then persisted and distributed to active members on the network.