CORE-13937: two party finality optimisation (#4268)
optimisation for two-party finality flow, where there is the possibility to skip second session communication by knowing the number of parties involved. There wouldn't be unseen signatures if there were two parties, so there is no need to send or receive those signatures. Instead, this implementation skips this part. 
In this PR I implemented finality flow version 2 and the payload will be a POJO containing `initialTransaction` and `transferAdditionalSignatures` instead of `initialTransaction` only in payload as in version 1. This POJO will be serialized and deserialized in the follow-up PR