CORE-15497: make use of notary passed from request body (#4334)
This PR changed api of registerStaticMember and context notary body in MGM registration for e2e testing, so that we can make use of notary that has passed to request body of demo flow, as opposed to notary not being used previously. Now the update will use either notary that was specified in call site or whatever notary that was available in network.