CORE-10114 Introduce custom metadata in static networks (#4330)
Allows members of a static network to specify custom ("ext.") properties in their registration context which will be included in their MemberInfo. Also allows properties (minimum platform version and custom "ext." parameters) to be specified in the group policy of static network CPIs which will be included in the group parameters of the network.