CORE-14232 - Upgrade json schema validator library (#4341)
Upgrading the json-schema-validator library to get rid of some spurious WARNING logs around keywords that are not identified (e.g. defs / then / else) even though they are supported by that version of the spec. Some more discussion and details available on this issue. I didn't upgrade up to the latest version of the spec (V202012) on purpose, because when I attempted it there were errors on some of our tests, e.g. ExternalChannelsConfigValidatorTest was failing expecting exceptions in case of unknown properties that weren't caught by validation. The release notes for that version of the spec do not provide any commentary on backwards compatibility in contrast to previous versions, so it's possible there are breaking changes. Given the scope of this is just to eliminate the warning, upgrading to V201909 should be sufficient. I also didn't upgrade to the latest version of json-schema-validator (1.0.86), because that depends on version 2.0.7 of SLF4j, whereas Corda uses version 1.7.36 which is the only one provided at the OSGi level, thus leading to OSGi resolution errors. I upgraded to the latest version of the library that depends on version 1.7.36 of SLF4j, while also containing the fix we need. I also removed the exclusion of com.fasterxml.jackson.core:jackson-databind as a transitive dependency of it in the :libs:serialisation module, since that has now been upgraded to the latest version that doesn't have a CVE.

I also had to force use of a specific version (1.7.3) of com.ethlo.time, because the one used by the library by default (1.7.0) does not have an OSGi bundle.