CORE-15182 - Added a check that ensures no contract state type has more than one UTXO token observer (#4359)
The verification is performed when the sandbox is created. This means the verification is only done during runtime and a flow can failed because more than one UTXO token observer was found. This is not ideal but can be improved in the future.

Refactored the code so the values of the container that maps a `token state` to `UTXO token observer` is a single instance and not a list.