CORE-15297: refactor crypto processor to reduce layering (#4369)
Combine the signing service with the crypto service, and remove more of the support for non-soft HSMs. That removes around 1200 lines of code, and address these tickets:

CORE-15276: remove the unused crypto throttling layer, which was never triggered. The architectural choice of Corda 5 is to retry from the top, not at each layer, and crypto is never at the top. Mutliple levels of retires are problematic.
CORE-15277: move the handling of exceptions out of a proxy layer into particular places where exceptions are expected. The main motivation was to reduce the number of places we have to touch on each interface change.
This required some movement of pieces to keep the package import graph from becoming cyclic with the above changes. In particular SigningKeyInfo and SigningKeySaveContext more into the persistence package, and TenantInfoService moves from crypto-service to crypto-softhsm-impl. We remove many test cases, which are testing layers that no longer exist, and move other test cases to reflect the newly simplified structure. All the tests in SigningServiceGeneralTests move to SoftCryptoServiceGeneralTests.

We also remove the hsmId and instance fields from CryptoServiceRef, since those are associated with layers that have been collapsed.