CORE-15351: log trace span to separate log file (#4299)
When deploying Corda with the combined worker, if the following command line argument is provided:  --send-trace-to=http://localhost:9411, the trace data (trace spans) is sent over to the ZipKin Server, as a valid host URL has been provided, however if no URL is provided the tracing data is not generated nor logged anywhere. With this change, when no ZipKin host URL is provided the trace spans will be logged locally into a file called corda-trace-spans.log. Developers have the flexibility to configure logging behaviour through the log4j2.xml file. They can choose to enable or disable this local logging with or without ZipKin and can also opt to direct logs to the console instead of log files if they prefer.