CORE-14784 Add e2e test config manager to update and revert configs per test. (#4363)
Added library to the e2e test utilities for testing with specific configurations. This provides common handling of configurations across e2e tests which includes reverting configuration. SingleClusterTestConfigManager can be use to manage the config of a specific or default single cluster, and MultiClusterTestConfigManager can be used to manage the same config across multiple clusters. Configuration overrides can be loaded for multiple properties and configuration sections before applying them to the cluster(s). The configuration manager implements AutoCloseable so these can be used with try-with-resources to ensure the configuration is reverted once no longer needed. Configuration will always be reverted back to it's original state before the test run.

A utility method is included called managedConfig which allows test writers to pass in a lambda which once complete will revert any test configuration set.

This also provides clear logging in the test output about what configuration values were set during a test run and confirmation that they were reverted. ClusterInfo was updated to expose a name property for clearer logging.

Two tests in this repo which were already testing with configuration changes have been updated to use the new library.