CORE-12005 - Adjust MGM onboarding command and update corresponding documentation (#4323)
This PR addresses CORE-12005 by adjusting the existing command for MGM onboarding. The changes involve modifications to command parameters and options, as well as introducing the ability to (over)write the newly created group ID to a file, to be used as the default with lookup, check-registration-status, and operate commands.