CORE-15285: Change sampling logic to favour flow traces (#4423)
The current configurability of traces is limited to a provided sample rate or unlimited sampling for all system requests, agnostic of their source or purpose. To reduce unnecessary data from being sampled, developers usually define a sample rate of their preference. However, this sample rate will be applied to all requests arbitrarily, meaning that the trace one is interested in might not get sampled.

When it comes to tracing as a whole, we are more interested in flow request sampling. With this change, we ensure that whatever sample rate is defined, is applied to all requests from URI endpoint /flow with method POST, as well as all other requests from the server. The change currently supports the following REST API versions: v5_1,, v1.