CORE-10598 Reconcile the members topic with the member info table (#4398)
Add a reconciler to reconcile the members topic with the member info table. The member info table is only used by an MGM Virtual Node and we don't support deletion. I have added the field is deleted to the table as this is required by the reconciler code. The reconciler will run on a non-MGM Virtual Node but this won't cause any functional issues because there won't be any records with isDeleted = true, so it won't write any tombstone records.

I have updated the suspend and activate member MGM rest commands to make the serial number mandatory (the old rest endpoints are deprecated for removal in a later version). This makes the API is less prone to user error, the MGM operator will have to check the members memberInfo for the serial number before suspending the member.

See also corda/corda-api #1132.