CORE-11990 - Modify signed data handling (#4425)
Review how we handle signed data for member info and registration request. We should store and retrieve the data in its original serialised form instead of serialising and deserialising back and forth.
Remove queryMembersSignatures and get the signature information for data distribution by using queryMemberInfo instead.

As part of this we had to modify the member info to contain signed information and the registration request to contain the original byte array along with the member signature.