CORE-16266: Allow the flow processor to run the fiber multiple times as part of the same event execution (#4477)
Adjust the flow pipeline to repeatedly evaluate the checkpoint after executing the user code and doing any post-processing on the outputs. In cases where the user code can immediately run again, this process will ensure that that happens without requiring an event to be published back to the flow event topic.

As part of this solution, the flow pipeline has been split into two classes. Longer term it is probably worth splitting the pre-processing step out into its own class and removing the full pipeline class entirely (instead chaining the classes for the individual steps in the FlowEventProcessorImpl), but that is out of scope for this PR. The new class is responsible for executing the user code, doing any processing of the checkpoint on output, and establishing whether the flow is eligible to be run again.

A change is also required to the flow runner to ensure that the decision to start a new flow or continue an existing one is made based on both the checkpoint state and input event. Without this the runner would repeatedly start the same flow if the original start event lets the user code execute past multiple suspension points.

Note that the integration tests have been temporarily disabled. Fixing these is likely to be a significant effort, and a future PR will also require integration changes. Therefore all integration test fixes will be carried out in a follow up PR.