CORE-16397 Remove reformatting of certificates and other types in CLI tool output (#4509)
The CLI tool was removing new lines to make output more readable but by doing so it altered certificates and so they could not be parsed. Certificate parsers expect new lines at certain points and so by removing the new lines the certificates were seen as invalid. Removed the output manipulation and added unit tests to prevent this occurring again. Refactored the print utils class also to be more concise.