CORE-15069 move flow to components (#4164)
Move two flow modules from libs to components. This will be used in future work to make use of a membership service in components to check clusters are local. Moving or copying this service was considered also but the move is also suggested for better design as these modules are closer to components than libs.