CORE-16266: Remove remaining flow Wakeup events (#4499)
Removes the instances where Wakeup events are published. This primarily involves changing request handlers, which are responsible for dealing with the request the user has made out of flow code. There are also some changes to test code to remove references to Wakeup events, which were frequently used as a placeholder type in the tests.

For retries, the input event is published back to the flow event topic to be processed again later. This mechanism should be sufficient to get past transient problems with running a flow. If this fails enough times, a fatal error will occur as before. Note that this requires no scheduling, and therefore no Wakeup event is required. All Wakeup events can therefore be removed from the system.

For testing, all unit and e2e tests are passing.