CORE-16612: Force quasar to ignore metrics packages when instrumenting code (#4522)
Adds the `@QuasarIgnoreAllPackages` annotation to the metrics library to avoid instrumenting this.

This avoids an issue where an exception occurs inside the metrics library indicating that an object is null, despite this not being possible from the compiler's perspective. This happens because Quasar is instrumenting the metrics library, and incorrectly restoring the metrics object on the stack after a suspension. There is no suspendable code inside the metrics library, and so this instrumentation is not required.