CORE-16228 Splitting FlowWorker into Flow and Verification workers (#4521)
The aim of this commit is to split out the VerificationProcessor from the FlowWorker into its own worker: a VerificationWorker. This will allow for more granular scaling of Corda clusters in response to increased traffic. This is part of a larger epic to create a more focused set of Corda workers.