CORE-16808: Add a Cloudwatch metrics registry (#4560)
Introduces a Cloudwatch metrics registry that is enabled if the environment supports it and the correct service account is applied.

Corda can provide a large volume of metrics data when run at scale. This can lead to challenges with running monitoring infrastructure. Addressing this requires both a review of the monitoring stack and of the data that Corda is exporting to ensure this is sensible. In the short term, Corda needs to integrate with a monitoring system that copes with the volume of data being exported, in order for us to do some initial testing.

This is a very rough solution to this problem. In particular, the mechanism for providing the AWS SDK in the OSGi framework is very blunt and does not follow the pattern we've established for working with other non-OSGi dependencies. In the short term this is probably ok. Longer term a more fully fledged metrics infrastructure is likely needed, which will remove the need for this change.
---------

Co-authored-by: Oscar Dalmau <oscar.dalmau@r3.com>