CORE-16901: Do not require a wakeup event when platform errors are encountered (#4583)
Allow platform errors to immediately wake up the flow if encountered, by catching them in the `FlowExecutionPipelineStage` and returning an error continuation immediately.

Platform errors are thrown in the flow pipeline if an error is encountered when calling an API, and are used as a signal that information about the failure should be returned to the user code. Previously when this happened, the exception triggered the flow pipeline to abort processing of the inbound event and record the error on the checkpoint. The flow was then woken up by the scheduler at some later point, which would trigger the platform error to be passed back to the user code.

By allowing platform errors to immediately return an error to the fiber, most cases of its use are handled. There are two exceptions:

- The CounterpartyFlowInfoWaitingForHandler can throw a platform exception. As a waiting for handler, the job of this class is to evaluate the current state of the checkpoint alongside the input event, and determine which flow continuation to return. In this case, the same condition can be handled by immediately returning a flow continuation instead, rather than throwing the excpetion.
- The FlowGlobalPostProcessor may throw a platform exception if the counterparty to a session does not exist. This is inherently dangerous, as the flow could wake up from calling any API (or even exiting a subflow) with an error for something entirely unrelated to what the flow was doing. However, the check must be done here as it is possible for a counterparty to cease to exist after the flow has requested a message be sent (for example if the member is suspended). Consequently, changing this to a fatal exception is the most consistent approach.