CORE-15391 Add support for custom checkpoint state (#4599)
CORE-15391 Add support for custom state to be stored in the checkpoint. This has been added to allow external flow event handlers to store local state that can span flow suspensions. As part of this change the state used for metrics has been migrated to use the new custom state.