CORE-16879 - use max retry window duration  (#4587)
Change transient flow exception handling to use a time window rather than a retry count.

Previously retries for transient flow issues were scheduled, meaning plenty of time was allowed for the underlying transient problem to clear. With the removal of the scheduler, it's possible for the transiently failing event to be replayed through to the retry limit very quickly in low load scenarios. Changing the limit to be time-based ensures that enough time passes before the flow is treated as failed.

---------

Co-authored-by: James Higgs <james.higgs@r3.com>