CORE-17024: Remove the non-serializable FlowCheckpointImpl from the fiber stack (#4612)
Ensure that FlowCheckpointImpl is not serialized in the fiber.

This class is marked as non-serializable, so accidentally serializing it results in flow failure. This could previously happen if the session was closed, due to a change in close logic. By moving the reference to this class out of the function stack at the serialization point, this error can be prevented.