CORE-17051: Ensure pipeline context is updated with intermediate flow results (#4620)
The flow pipeline context is used when an error occurs to signal what the problem was back to the user. In cases where the flow has only just started, it might not be possible to use this to create a new status. Previously, this condition held even if the pipeline has executed through multiple stages of the flow fiber. This context should now be updated with intermediate values, ensuring they can be used if an error does occur.

This change also updates a log message to be less misleading in the event that a flow status cannot be published.