CORE-16319: State Manager API & Implementation (#4602)
Introduce the StateManager API and a postgres db implementation.

The State Manager is a new component responsible for retrieving and updating a generic state from an out-of-process storage. It is entirely agnostic of the actual state stored. Every state stored within the underlying persistent storage will have an associated metadata, modifiedTime and version.

This change is responsible for adding the libraries to the repository and do not deploy and configure the state manager in any workers. This is the goal of CORE-16318.

Co-authored-by: Conal Smith <conal.smith@r3.com>
Co-authored-by: Juan José Ramos Cassella <jujoramos@gmail.com>