CORE-17065: Ensure session messages get routed in the flow mapper correctly (#4630)
Restructures the `RecordFactoryImpl` class in the flow mapper impl to better divide responsibilities in this module. This class is now responsible for generating the correct forwarding records, while the other classes are responsible for identifying when a session record needs to be forwarded.

This fixes two issues:
 - The destination identity was mis-identified when trying to establish whether an identity was local to the cluster, which causes a bug in multi-cluster scenarios.
 - The sequence number was washed off in some cases when generating a session record to forward to the P2P layer.

This change also introduces a better log in the case of transient flow exceptions to identify what the flow is waiting for.