CORE-17115: Send error messages back to the counterparty when data is received and the mapper state is null (#4650)
Adds the ability to send a session error event back to the counterparty if the session is unknown when receiving a data event.

The flow mapper could receive a session data event with no session init information included. Ordinarily this is a normal part of the flow message exchange. However, if the session is unknown to the flow mapper then this is an error, as the mapper cannot route the message successfully. Previously the mapper would attempt to forward these messages anyway and fail. With this change, the mapper will return an error to the counterparty for inbound messages, and drop the event with a warning for outbound ones.