CORE-10087 - Introduce utils needed to migrate network tests to the e2e-tests repo (#4654)
Introduce utilities needed for e2e testing in order to migrate some e2e tests into the e2e-tests repo (MultiClusterDynamicNetworkTest and SessionCertificateTest). As part of this, I extracted disabling the certificate revocation checks from the onboardMember / onboardMgm functions, so that it can be executed along with any other config changes needed within the test. Otherwise, this could revert any changes on the same section performed by within the test (e.g. when trying to change tls mode, which is in the same gateway section as revocation checks). I also removed usage of any proxy utils that are targeted for removal/deprecation from e2e testing and stuck to using the existing utilities (this meant I had to introduce a couple of new functions that were not available).