CORE-16860: Send counterparty explicit requests for flow version info when api is called. (#4676)
Send a counterparty info request to start a session if no data is sent before requesting counterparty info.

It is possible when starting a session with a counterparty to request counterparty information before sending the first data message. Previously if this happened with the new session protocol for 5.1, the flow would fail. This PR introduces specific messages to handle this case in the counterparty info request and response message types. These will now be sent if this condition is met.

---------

Co-authored-by: James Higgs <james.higgs@r3.com>