CORE-12411 - Add declination reason (#4696)
This change adjusts the MGM code, so that it communicates the reason for declined requests back to the user and the appropriate handler, so that this information is stored in the DB and surfaced to the user. In some cases, we don't want to reveal potentially sensitive information to a member (especially in cases where the request hasn't gone fully through authentication yet). In these cases, a general message is communicated instructing the user to communicate with the network operator for more details.