CORE-17354 Force initial reconciliation to support updating configuration after upgrade (#4708)
Before this change, If we extended the config schema e.g. add a property X, and we upgraded a cluster we would had expected property X to be available to the components that need it. However, that was not the case, property X was not available and the components trying to access it threw on trying to get it from provided config.

What happens is, reconciliation compares DB records to Kafka records (based on version field) and the ones found to be needing updating on Kafka get published to Kafka. But right before getting published, they go through the defaulting process where properties not present in DB config get added with defaulted values.

In the above scenario, Kafka is already populated (since we upgrade), meaning Kafka config version match DB config version and therefore no update is needed. Meaning we also missed going through the defaulting process and that's where we missed the newly added property X.

- Forces the initial reconciliation (only) to do a full reconciliation (even for DB records that are of same version compared to Kafka ones) to take all DB records through the defaulting process and so in case of updated config schemas, the updated config schemas will be picked up and the new properties will be added through the defaulting process, and then published to Kafka