CORE-17360: Update execute cleanup event to match new avro object (#4719)
Updates usages of the ExecuteCleanup Avro object to include an empty list of IDs.

Currently the IDs field is unused, but this will be taken advantage of in the new version of the flow mapper cleanup code. The old code still uses these events however, and so needs to be kept in line with the new event definition.